Marty : Refactoring code mission by commenting the call of a not used script named WASP\Init_Client.sqf under initJIPCompatible.sqf.
Refactoring code mission by commenting the call of a not used script named WASP\Init_Client.sqf under initJIPCompatible.sqf.

This script was used to check if the player object was not null. It was checked in a loop using the waitUntil sqf command causing unnecessary resources consumption.
It was used into another script called car_wheel_new.sqf to change the wheel of vehicle in the action menu, which is not used in the mission.

To do:
-	First step, comment lines in initJIPCompatible.sqf where the WASP\Init_Client is executed and monitor for few days if it’s not impacting the mission script generating weird bug.
-	After few days, remove the commented line above.
-	Delete the WASP\Init_Client.sqf file.
-	Delete the car_wheel_new.sqf file.
-	Remove the commented lines in addAction.sqf