HQ wreck mark is updated on map in case the wreck is moved.
It's a very rare situation that can happen when a player lift the HQ wreck with helicopter to repair it in a different location. In that case it is good to display the location of the HQ wreck in case the helicopter crash during lifiting so HQ wreck location is ALWAYS known to the friendly team.