Add seperate upgrade levels, placeholder text, guessed by the LLM, expect IR smoke which was in the image
Modify this text to have all of the upgrade levels that are on the image you received. Do not add levels yet, only the upgrade names as H2. Be sure to write a short description of each of the upgrade levels, you will try to guess what each of the upgrades do (this is a military simulator game).