Add the levels as bullet points from the image data
Read the image for each of the upgrades. Figure out how many levels each of the upgrade has first, and then provide the output. For example, the Aircraft Factory is 4/5 so it means that it's current level is 4, and maximum level is 5. BE SURE TO REPEAT THIS SAME LOGIC FOR REST OF THE LEVELS. Disregard the current level, and only make as many H3's for this text as there are maximum levels. Instead of guessing what each of the levels do, just insert [INSERT LEVEL DESCRIPTION HERE] to each of the levels with a bullet point. This is in a context of military simulation game!!

Text to edit and provide as output in ```: