Marty : Bomb Dropping Restriction Beyond Altitude Limit.
This update introduces a restriction on dropping all types of bombs known to date on the server, at a maximum altitude set by default to 2500 meters. This measure aims to prevent players from exploiting the 3000-meter limit of surface-to-air missile fire.

If a player drops a bomb beyond the regulated altitude, the bomb will automatically disintegrate and the occupants of the aircraft will receive a notification via the vehicle channel. The notification message supports multiple languages and is available in 6 different languages.

Additionally, administrators can adjust the altitude restriction parameter in the launch settings.

Finally, the code has been optimized using an event handler only for flying vehicles.