Remove walls from the rest of the sides of the factories, so they will have 4 entrances in total.
For Barracks, Light, Heavy and Aircraft factory (Ignore COMMANDCENTER):
All of these are factories that have walls built on the side of them. However, on one side, the wall is missing from the center. Explain your reasoning to identify this hole and replicate it for the all of the other three sides, commenting out the line.